move `getSchemaRegistryWorkflow()` out of commands/docker for easier testability ... and also sanity
