+++ src/docker/workflows/index.ts
@@ -1,8 +1,9 @@
 import net from "net";
 import { window } from "vscode";
-import { getLocalKafkaImageName } from "../configs";
+import { getLocalKafkaImageName, getLocalSchemaRegistryImageName } from "../configs";
 import { LocalResourceWorkflow } from "./base";
 import { ConfluentLocalWorkflow } from "./confluent-local";
+import { ConfluentPlatformSchemaRegistryWorkflow } from "./cp-schema-registry";
 
 // maybe this can live somewhere else if we need it for more than just container creation:
 /** Look for an available port on the host machine and return it. */
@@ -32,6 +33,20 @@ export function getKafkaWorkflow(): LocalResourceWorkflow {
       window.showErrorMessage(errorMsg);
       throw new Error(errorMsg);
     }
+  }
+  return workflow;
+}
+
+/** Determine which Schema Registry workflow to use based on the user-selected configuration. */
+export function getSchemaRegistryWorkflow(): LocalResourceWorkflow {
+  const imageRepo: string = getLocalSchemaRegistryImageName();
+  let workflow: LocalResourceWorkflow;
+  switch (imageRepo) {
+    case ConfluentPlatformSchemaRegistryWorkflow.imageRepo:
+      workflow = ConfluentPlatformSchemaRegistryWorkflow.getInstance();
+      break;
+    default:
+      throw new Error(`Unsupported Schema Registry image repo: ${imageRepo}`);
   }
   return workflow;
 }
